Stage one pi compatibility
Worked out installation, and I should in theory have that up and
running. Fucked up on which arm version was selected for node, but
that's a quickfix.

Haven't tested activation because I'm dumb, forgot which version I just
installed, and the lack of @latest symlinks/tracking is starting to
annoy me.

Should also start tracking which versions are actually installed. Not
sure if that's better suited for metadata, or if that's something
I'll just compute at runtime. Not particularly performant, especially if
the underlying file system is shit, but that's a problem for me to
figure out when I start diving into this rabbit hole.