Add renderdoc, fix symlink bug
Symlinking was a bit too lenient with directory creation. The previous
system symlinked any non-existent directories, under the assumption that
all the defaults totally definitely existed, which doesn't work in
practice.

The new check checks for directories matching the name of the package,
which isn't perfect either, but it's more likely to result in a bit too
many symlinks (such as in the case of node, where the package is called
nodejs but the directories are called node). It will not result in
directory conflicts.