Migrate git to stc::Unix::Process
This commit also nukes the shell arg escape function. It wasn't used as
wildely as it should, so there's still injection sites, but more
importantly, the function is made completely redundant by
stc::Unix::Process. No shell is needed in the vast majority of cases.
Only one of the commands migrated so far has actually needed it, and
it's just a convenience because it spawns two extra git calls, so the
alternative is expanding it to three calls with some input preprocessing
due to trailing newlines. That'll stay a shell for now, but the addition
of working directories via an alternate route that doesn't need a shell
command to do it means the command doesn't take any user input, so it'll
never execute anything.

There's still plenty of ways user input can be abused, but there's no
more shell injection to worry about.